Fix: Broke out gallery nav modal into separate component
